@@ -31,7 +30,7 @@ classifiers = [
     "Topic :: Software Development :: Libraries :: Application Frameworks"
 ]
 readme = "README.md"
-requires-python = ">=3.10,<3.13"
+requires-python = ">=3.11,<3.13"
 
 # SPDX licence expression + explicit licence file (PEP 639)
 license = "Apache-2.0"
